+++ src/Traits/HasPermissions.php
@@ -2,22 +2,38 @@
 
 namespace Spatie\Permission\Traits;
 
-use Spatie\Permission\PermissionRegistrar;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use Illuminate\Database\Eloquent\Relations\Relation;
 use Spatie\Permission\Contracts\Permission;
+use Spatie\Permission\Contracts\Restrictable;
 use Spatie\Permission\Exceptions\GuardDoesNotMatch;
+use Spatie\Permission\PermissionRegistrar;
 
 trait HasPermissions
 {
     /**
+     * A model may have multiple direct permissions.
+     * @return BelongsToMany|Relation
+     */
+    abstract public function permissions(): BelongsToMany;
+
+    /**
      * Grant the given permission(s) to a role.
+     * If a restrictable instance is given, given permission(s) is/are scoped to it,
+     *  otherwise there won't be a scope for the permission(s).
      *
      * @param string|array|\Spatie\Permission\Contracts\Permission|\Illuminate\Support\Collection $permissions
-     *
+     * @param \Spatie\Permission\Contracts\Restrictable $restrictable
      * @return $this
      */
-    public function givePermissionTo(...$permissions)
+    public function givePermissionTo($permissions, Restrictable $restrictable = null)
     {
-        $permissions = collect($permissions)
+        // Permission objects, if directly collected, becomes arrays of fields and the flatten() messes with
+        // the map function giving every single Permission field as parameter for getStoredPermission.
+        // To avoid this, if a Permission is given an empty collection is created and the permission is pushed inside.
+        // In this way, in case of a Permission instance, the object is not flattened,
+        // but for arrays, collections and string everything works as expected.
+        $permissions = (($permissions instanceof Permission) ? collect()->push($permissions) : collect($permissions))
             ->flatten()
             ->map(function ($permission) {
                 return $this->getStoredPermission($permission);
@@ -25,9 +41,19 @@ trait HasPermissions
             ->each(function ($permission) {
                 $this->ensureGuardIsEqual($permission);
             })
+            // Attach takes ids, we retrieve them
+            ->map(function ($permission) {
+                return $permission->id;
+            })
             ->all();
 
-        $this->permissions()->saveMany($permissions);
+        // If there is no restrictable instance, we won't add anything on the pivot table,
+        //  which will default to null values on the restrictable morph.
+        // Otherwise we set the references to it
+        $this->permissions()->attach($permissions, is_null($restrictable) ? [] : [
+            'restrictable_id' => $restrictable->getRestrictableId(),
+            'restrictable_type' => $restrictable->getRestrictableTable(),
+        ]);
 
         $this->forgetCachedPermissions();
 
@@ -35,29 +61,31 @@ trait HasPermissions
     }
 
     /**
-     * Remove all current permissions and set the given ones.
+     * Remove all current not scoped permissions and set the given ones.
+     * If a Restrictable instance is given, permissions will be removed and set only for that instance scope.
      *
      * @param string|array|\Spatie\Permission\Contracts\Permission|\Illuminate\Support\Collection $permissions
-     *
+     * @param \Spatie\Permission\Contracts\Restrictable $restrictable
      * @return $this
      */
-    public function syncPermissions(...$permissions)
+    public function syncPermissions($permissions, Restrictable $restrictable = null)
     {
-        $this->permissions()->detach();
+        $this->permissions($permissions)->detach();
 
-        return $this->givePermissionTo($permissions);
+        return $this->givePermissionTo($permissions, $restrictable);
     }
 
     /**
      * Revoke the given permission.
+     * If a Restrictable instance is given, the permission will be removed only for that resource scope.
      *
      * @param \Spatie\Permission\Contracts\Permission|string $permission
-     *
+     * @param \Spatie\Permission\Contracts\Restrictable $restrictable
      * @return $this
      */
-    public function revokePermissionTo($permission)
+    public function revokePermissionTo($permission, Restrictable $restrictable = null)
     {
-        $this->permissions()->detach($this->getStoredPermission($permission));
+        $this->permissions($restrictable)->detach($this->getStoredPermission($permission)->id);
 
         $this->forgetCachedPermissions();
 
@@ -66,7 +94,6 @@ trait HasPermissions
 
     /**
      * @param string|array|\Spatie\Permission\Contracts\Permission|\Illuminate\Support\Collection $permissions
-     *
      * @return \Spatie\Permission\Contracts\Permission
      */
     protected function getStoredPermission($permissions): Permission
@@ -88,7 +115,7 @@ trait HasPermissions
     /**
      * @param \Spatie\Permission\Contracts\Permission|\Spatie\Permission\Contracts\Role $roleOrPermission
      *
-     * @throws \Spatie\Permission\Exceptions\GuardMismatch
+     * @throws \Spatie\Permission\Exceptions\GuardDoesNotMatch
      */
     protected function ensureGuardIsEqual($roleOrPermission)
     {
@@ -102,7 +129,7 @@ trait HasPermissions
      */
     protected function isAssignedToGuard(): bool
     {
-        return (bool) $this->getGuardName();
+        return (bool)$this->getGuardName();
     }
 
     /**