+++ tests/resources/views/scopedHasAllRoles.blade.php
@@ -1,0 +1,5 @@
+@hasallroles($roles, null, $class, $id)
+does have all of the given scoped roles
+@else
+does not have all of the given scoped roles
+@endhasallroles