+++ tests/resources/views/scopedHasAnyRole.blade.php
@@ -1,0 +1,5 @@
+@hasanyrole($roles, null, $class, $id)
+does have some of the scoped roles
+@else
+does not have any of the given scoped roles
+@endhasanyrole