+++ tests/resources/views/scopedHasRole.blade.php
@@ -1,0 +1,5 @@
+@hasrole($role, null, $class, $id)
+has scoped role
+@else
+does not have scoped role
+@endhasrole