+++ tests/resources/views/scopedRole.blade.php
@@ -1,0 +1,5 @@
+@role($role, null, $class, $id)
+has scoped role
+@else
+does not have scoped role
+@endrole